requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ndg-httpsclient==0.5.1
pusher==3.3.3
pyasn1==0.6.1
pycparser==2.22
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
